LVPN-10909: Move openURI to tray instance
